Deep-EIoU/reid/requirements.txt
numpy
loguru
lap
Cython
h5py
Pillow
six
scipy
opencv-python
matplotlib
tb-nightly
future
yacs
gdown
flake8
yapf
isort==4.3.21
imageio
chardet
tabulate
pycocotools
thop
# Export --------------------------------------
# onnx
# onnx-simplified
# openvino-dev
# openvino2tensorflow
# tensorflow
# tensorflow_datasets